+++ components/tcf/firstLayer/src/index.js
@@ -1,106 +1,180 @@
-import React, {useState, useEffect} from 'react'
+import React, {useState, useEffect, useRef, useCallback} from 'react'
 import PropTypes from 'prop-types'
 import SuiButton from '@s-ui/react-atom-button'
 import SuiModal from '@s-ui/react-molecule-modal'
-import IconClose from './components/IconClose'
-import InfoCard from './components/InfoCard'
+import SuiNotification from '@s-ui/react-molecule-notification'
+import IconClose from './iconClose'
+import {I18N} from './settings'
+
+const CLASS = 'sui-TcfFirstLayer'
+// Adevinta rule of 250px of scroll to directly accept consent
+const SCROLL_TO_ACCEPT = 250
 
 export default function TcfFirstLayer({
-  loadUserConsent,
+  lang = 'es',
+  logo,
+  isMobile,
+  isTablet,
   getVendorList,
   saveUserConsent,
-  openSecondLayer
+  openSecondLayer,
+  openCookiepolicyLayer
 }) {
-  const [saveButtonDisabled, setSaveButtonDisabled] = useState(true)
-  const [state, setState] = useState(null)
-  const [vendorListState, setVendorListState] = useState(null)
-  const [modalOpen, setModalOpen] = useState(true)
+  const [show, setShow] = useState(true)
+  const i18n = I18N[lang]
 
+  const handleCookiePolicyLayerClick = useCallback(() => {
+    cookiesPolicyLink.current &&
+      cookiesPolicyLink.current.removeEventListener(
+        'click',
+        handleCookiePolicyLayerClick
+      )
+    openCookiepolicyLayer()
+  })
+  const cookiesPolicyLink = useRef()
+  const ref = useRef(null)
+  const textRef = useCallback(
+    node => {
+      if (ref.current) {
+        ref.current = null
+      }
+
+      if (node) {
+        cookiesPolicyLink.current = node.querySelector(
+          `.${CLASS}-body-info-link`
+        )
+        if (cookiesPolicyLink) {
+          cookiesPolicyLink.current.addEventListener(
+            'click',
+            handleCookiePolicyLayerClick
+          )
+        }
+        ref.current = node
+      }
+    },
+    [handleCookiePolicyLayerClick]
+  )
+
+  let initialYOffset
+  const checkScroll = () => {
+    if (
+      initialYOffset - window.pageYOffset >= SCROLL_TO_ACCEPT ||
+      initialYOffset - window.pageYOffset <= -SCROLL_TO_ACCEPT
+    ) {
+      handleSaveExitClick()
+    }
+  }
   useEffect(() => {
-    const loadConsent = async () => {
-      const {purpose} = await loadUserConsent()
-      setState({purposes: purpose})
-    }
-    const loadVendorList = async () => {
-      const {purposes} = await getVendorList()
-      setVendorListState({purposes})
-    }
-    loadConsent()
-    loadVendorList()
-  }, [getVendorList, loadUserConsent])
+    initialYOffset = window.pageYOffset
+    document.addEventListener('scroll', checkScroll, {passive: true})
+  }, [])
 
   const handleSettingsClick = () => {
     openSecondLayer()
   }
-  const formatConsentObject = ({purpose = {}}) => {
-    purpose.consents = purpose.consents || {}
-    purpose.legitimateInterests = purpose.legitimateInterests || {}
-    Object.keys(vendorListState.purposes).forEach(key => {
-      purpose.consents[key] = !!purpose.consents[key]
-      purpose.legitimateInterests[key] = !!purpose.legitimateInterests[key]
-    })
-    return purpose
-  }
-  const handleSaveExitClick = () => {
-    saveUserConsent(
-      formatConsentObject({
-        purpose: state.purposes
-      })
-    )
-    setModalOpen(false)
-  }
-  const handleAcceptAll = () => {
-    setState(prevState => {
-      const {purposes} = prevState
-      const {consents} = purposes
-      for (const key in vendorListState.purposes) {
-        consents[key] = true
-      }
-      return {
-        purposes: {consents}
-      }
-    })
-    setSaveButtonDisabled(false)
-  }
-  const handlePurposesChange = ({index, value, name}) => {
-    setSaveButtonDisabled(false)
-    setState(prevState => {
-      const {consents} = prevState[name]
-      consents[index] = !value
-      return {[name]: {consents}}
-    })
+
+  const handleSaveExitClick = async () => {
+    const {
+      purposes: VLPurposes,
+      vendors: VLVendors,
+      specialFeatures: VLSpecialFeatures
+    } = await getVendorList({language: lang})
+
+    const userConsent = {
+      purpose: {
+        consents: {},
+        legitimateInterests: {}
+      },
+      vendor: {
+        consents: {},
+        legitimateInterests: {}
+      },
+      specialFeatures: {}
+    }
+
+    const {purpose, vendor, specialFeatures} = userConsent
+    for (const key in VLPurposes) {
+      purpose.consents[key] = true
+      purpose.legitimateInterests[key] = true
+    }
+    for (const key in VLVendors) {
+      vendor.consents[key] = true
+      vendor.legitimateInterests[key] = true
+    }
+    for (const key in VLSpecialFeatures) {
+      specialFeatures[key] = true
+    }
+
+    saveUserConsent(userConsent)
+    handleCloseModal()
   }
 
   const handleCloseModal = () => {
-    setModalOpen(false)
+    document.removeEventListener('scroll', checkScroll)
+    setShow(false)
   }
+
+  const Content = ({modalType}) => {
+    const baseBodyClass = `${CLASS}-body`
+    const bodyClasses = isTablet
+      ? `${baseBodyClass} ${baseBodyClass}--footer ${baseBodyClass}--tablet`
+      : `${baseBodyClass} ${baseBodyClass}--footer`
+    const baseButtonsClass = `${CLASS}-buttons`
+    const buttonsClasses = isTablet
+      ? `${baseButtonsClass} ${baseButtonsClass}--footer ${baseButtonsClass}--tablet`
+      : `${baseButtonsClass} ${baseButtonsClass}--footer`
+
+    return (
+      <div className={modalType ? `${baseBodyClass}` : bodyClasses}>
+        <span
+          className={`${baseBodyClass}-info`}
+          ref={textRef}
+          dangerouslySetInnerHTML={{__html: i18n.BODY}}
+        />
+        <div className={modalType ? `${baseButtonsClass}` : buttonsClasses}>
+          <SuiButton onClick={handleSettingsClick} design="outline">
+            {i18n.CONFIGURE_BUTTON}
+          </SuiButton>
+          <SuiButton onClick={handleSaveExitClick}>
+            {i18n.CONTINUE_NAVIGATION_BUTTON}
+          </SuiButton>
+        </div>
+      </div>
+    )
+  }
+  Content.propTypes = {
+    modalType: PropTypes.bool
+  }
+
   return (
-    <div className="sui-TcfFirstLayer">
-      <SuiModal
-        isOpen={modalOpen}
-        closeOnOutsideClick
-        closeOnEscKeyDown
-        header={<strong>TcfFirstLayer</strong>}
-        iconClose={<IconClose />}
-        onClose={handleCloseModal}
-      >
-        {state && vendorListState && (
-          <InfoCard
-            name="purposes"
-            title="We care about your privacy"
-            descriptions={vendorListState.purposes}
-            state={state.purposes}
-            onStateChange={handlePurposesChange}
-          />
-        )}
-        <br />
-        <br />
-        <SuiButton onClick={handleSettingsClick}>Manage Settings</SuiButton>
-        <SuiButton onClick={handleSaveExitClick} disabled={saveButtonDisabled}>
-          Save and Exit
-        </SuiButton>
-        <SuiButton onClick={handleAcceptAll}>Accept All</SuiButton>
-      </SuiModal>
+    <div className={CLASS}>
+      {isMobile && (
+        <SuiModal
+          isOpen={show}
+          closeOnOutsideClick
+          closeOnEscKeyDown
+          header={<img className={`${CLASS}-logo`} src={logo} alt="logo" />}
+          iconClose={<IconClose class={`${CLASS}-icon-close`} />}
+          onClose={handleCloseModal}
+          fitContent
+        >
+          <Content modalType />
+        </SuiModal>
+      )}
+      {!isMobile && (
+        <div className={`${CLASS}-notification`}>
+          <SuiNotification
+            position="bottom"
+            autoClose="manual"
+            show={show}
+            showCloseButton={false}
+            variation="positive"
+            type="system"
+          >
+            <Content />
+          </SuiNotification>
+        </div>
+      )}
     </div>
   )
 }
@@ -108,7 +182,11 @@ export default function TcfFirstLayer({
 TcfFirstLayer.displayName = 'TcfFirstLayer'
 TcfFirstLayer.propTypes = {
   openSecondLayer: PropTypes.func,
-  loadUserConsent: PropTypes.func,
+  openCookiepolicyLayer: PropTypes.func,
   getVendorList: PropTypes.func,
-  saveUserConsent: PropTypes.func
+  saveUserConsent: PropTypes.func,
+  isMobile: PropTypes.bool,
+  isTablet: PropTypes.bool,
+  lang: PropTypes.string,
+  logo: PropTypes.string
 }