+++ demo/tcf/utils/borosMock.js
@@ -222,6 +222,24 @@ const mockedVendorList = {
       specialFeatures: [],
       policyUrl: 'https://www.conversantmedia.eu/legal/privacy-policy'
     }
+  },
+  specialFeatures: {
+    '1': {
+      id: 1,
+      name: 'Use precise geolocation data',
+      description:
+        'Your precise geolocation data can be used in support of one or more purposes. This means your location can be accurate to within several meters.',
+      descriptionLegal:
+        'Vendors can:\n* Collect and process precise geolocation data in support of one or more purposes.\nN.B. Precise geolocation means that there are no restrictions on the precision of a user’s location; this can be accurate to within several meters.'
+    },
+    '2': {
+      id: 2,
+      name: 'Actively scan device characteristics for identification',
+      description:
+        "Your device can be identified based on a scan of your device's unique combination of characteristics.",
+      descriptionLegal:
+        'Vendors can:\n* Create an identifier using data collected via actively scanning a device for specific characteristics, e.g. installed fonts or screen resolution.\n* Use such an identifier to re-identify a device.'
+    }
   }
 }